api/requirements/base.txt
django==4.2
djangorestframework==3.15.1
django-taggit==6.0.0
django-filter==24.1
django-phonenumber-field==8.0.0
phonenumbers==8.13.30
django-celery-email==3.0.0
django-celery-beat==2.6.0
python-dotenv==1.0.1
djoser==2.3.0
drf-yasg==1.21.8
argon2-cffi==23.1.0
celery==5.4.0
redis==5.0.3
flower==2.0.1
python-magic==0.4.27
python-dateutil==2.9.0